requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.0.1
openai==1.30.1
transformers==4.40.2
torch==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu